upgraded cvm registry, chainid in signed message must be uint256

+++ contracts/src/usecases/bases/CVMSignature.sol
@@ -18,6 +18,6 @@ abstract contract CVMSignature {
         virtual
         returns (bytes memory)
     {
-        return abi.encodePacked(bytes(prefix), uint16(block.chainid), address(this), userData);
+        return abi.encodePacked(bytes(prefix), block.chainid, address(this), userData);
     }
 }
